✨ tweak: refactor logging system into modular architecture
